Add files via upload
In this project, we put everything together we've learnt about algorithms complexity. We will imagine that we own an online laptop store and want to build a way to answer a few different business questions about our inventory.

We will use the laptops.csv (https://bit.ly/3G32iYj) file as our inventory.

ID: A unique identifier for the laptop.
Company: The name of the company that produces the laptop.
Product: The name of the laptop.
TypeName: The type of laptop.
Inches: The size of the screen in inches.
ScreenResolution: The resolution of the screen.
CPU: The laptop CPU.
RAM: The amount of RAM in the laptop.
Memory: The size of the hard drive.
GPU: The graphics card name.
OpSys: The name of the operating system.
Weight: The laptop weight.
Price: The price of the laptop.